shovel the business logic into a standalone script
